build: add CFLAGS/LDFLAGS/LDLIBS, fix GCC compatibility issues
  Add CFLAGS, LDFLAGS, and LDLIBS to the build system. CFLAGS are
  passed to the C compiler, LDFLAGS before objects in the link step,
  and LDLIBS after objects and libc.a. This allows linking libgcc.a
  or passing compiler-specific flags when using a cross-toolchain.

  Split Makefile environment into HOST_BUILD_ENV and CROSS_BUILD_ENV.
  Host tool recipes use HOST_BUILD_ENV which resets all cross flags
  to prevent leakage. Add HOSTAS variable for consistency with HOSTCC.
  Simplify CC/AS/LD/AR defaults from $(origin) conditionals to plain
  assignments since command-line variables always override file variables.

  Fix GCC cross-compilation issues:
  - Add GNUC guard to libc stdarg.h so GCC uses its own va_arg
  builtins instead of the bootstrapped compiler's intrinsics
  - Change kernel string.h/string.c size parameters from unsigned int
  to unsigned long to match the standard signatures on AArch64
  - Add missing putchar() to libc (GCC optimizes printf to putchar)
  - Fix test_codegen.c for C11 compliance: add explicit pointer casts
  for incompatible pointer type assignments, use unsigned arithmetic
  for intentional overflow test, rename duplicate struct member,
  remove non-standard comma-operator-as-lvalue test

+++ kernel/string.c
@@ -1,7 +1,7 @@
 #include "string.h"
 
-unsigned int strlen(const char *s) {
-	unsigned int n = 0;
+unsigned long strlen(const char *s) {
+	unsigned long n = 0;
 	while (s[n]) {
 		n++;
 	}
@@ -16,7 +16,7 @@ int strcmp(const char *s1, const char *s2) {
 	return (unsigned char)*s1 - (unsigned char)*s2;
 }
 
-int strncmp(const char *s1, const char *s2, unsigned int n) {
+int strncmp(const char *s1, const char *s2, unsigned long n) {
 	while (n > 0 && *s1 && *s1 == *s2) {
 		s1++;
 		s2++;
@@ -28,7 +28,7 @@ int strncmp(const char *s1, const char *s2, unsigned int n) {
 	return (unsigned char)*s1 - (unsigned char)*s2;
 }
 
-char *strncpy(char *dst, const char *src, unsigned int n) {
+char *strncpy(char *dst, const char *src, unsigned long n) {
 	char *d = dst;
 	while (n > 0 && *src) {
 		*d++ = *src++;
@@ -41,7 +41,7 @@ char *strncpy(char *dst, const char *src, unsigned int n) {
 	return dst;
 }
 
-void *memmove(void *dst, const void *src, unsigned int n) {
+void *memmove(void *dst, const void *src, unsigned long n) {
 	char *d = dst;
 	const char *s = src;
 
@@ -59,7 +59,7 @@ void *memmove(void *dst, const void *src, unsigned int n) {
 	return dst;
 }
 
-void *memset(void *s, int c, unsigned int n) {
+void *memset(void *s, int c, unsigned long n) {
 	unsigned char *p = s;
 	while (n-- > 0) {
 		*p++ = (unsigned char)c;
@@ -67,7 +67,7 @@ void *memset(void *s, int c, unsigned int n) {
 	return s;
 }
 
-void *memcpy(void *dst, const void *src, unsigned int n) {
+void *memcpy(void *dst, const void *src, unsigned long n) {
 	char *d = dst;
 	const char *s = src;
 	while (n-- > 0) {
